+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2025.03.04
+
+### Changed
+
+- Generated etags are now seven characters long.
+- `Auto Tag` command is now `Toggle Auto Tag`.
+
 ## [0.0.5] - 2025.03.04
 
 ### Added